% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology_functions.R
\name{calculate_thresholds}
\alias{calculate_thresholds}
\title{Calculate multiple threshold methods for topology analysis}
\usage{
calculate_thresholds(data)
}
\arguments{
\item{data}{Numeric vector to calculate thresholds for}
}
\value{
A list of threshold values using different methods
}
\description{
Calculate multiple threshold methods for topology analysis
}
