% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology_functions.R
\name{is_topology_connected_manual}
\alias{is_topology_connected_manual}
\title{Check if a topology is connected using manual checking approach}
\usage{
is_topology_connected_manual(topology)
}
\arguments{
\item{topology}{A list of sets representing the topology}
}
\value{
logical TRUE if topology is connected, FALSE otherwise
}
\description{
Check if a topology is connected using manual checking approach
}
\examples{
topology <- list(c(1,2,3), c(3,4,5))
is_topology_connected_manual(topology)
}
