% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology_functions.R
\name{simplest_topology}
\alias{simplest_topology}
\title{Create a completely connected graph topology}
\usage{
simplest_topology(datos)
}
\arguments{
\item{datos}{Numeric vector containing the data points to analyze}
}
\value{
A list containing:
\itemize{
\item R - A list of relationships between vertices
\item subbase - The initial subbase of the topology
\item base - The base of the topology
\item topology - The final topology structure
}
}
\description{
Create a completely connected graph topology
}
\examples{
data <- c(1, 2, 3, 4, 5)
result <- simplest_topology(data)
}
