% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologyR-package.R
\docType{package}
\name{topologyR-package}
\alias{topologyR}
\alias{topologyR-package}
\title{topologyR: Topological Connectivity Analysis for Numeric Data}
\description{
Topological data analysis methods based on graph-theoretic approaches for discovering topological structures in data. Constructs topological spaces from graphs following Nada et al. (2018) \doi{10.1002/mma.4726}, with visibility graph construction for time series following Lacasa et al. (2008) \doi{10.1073/pnas.0709247105}. Supports directed visibility graphs for bitopological analysis of temporal irreversibility (Kelly, 1963), and Alexandrov topology construction from reachability preorders.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IsadoreNabi/topologyR}
  \item Report bugs at \url{https://github.com/IsadoreNabi/topologyR/issues}
}

}
\author{
\strong{Maintainer}: José Mauricio Gómez Julián \email{isadore.nabi@pm.me}

Authors:
\itemize{
  \item José Mauricio Gómez Julián \email{isadore.nabi@pm.me}
}

}
\keyword{internal}
